trying to implememnet the scaling, not sure whether this gonna work

+++ src/attributes/Memory.java
@@ -2,7 +2,7 @@ package attributes;
 
 public class Memory extends DoubleAttribute{
 
-	public Memory(int value){
+	public Memory(long value){
 		super(value);
 	}
