+++ src/protocols/BrandsAndChaumProtocol.java
@@ -44,7 +44,7 @@ public class BrandsAndChaumProtocol extends DBProtocol{
 		return false;
 	}
 	@Override
-	public int getMemory(int n) {
+	public long getMemory(int n) {
 		return getTotalBitsExchanged(n) + 2*n;
 	}
