+++ src/protocols/BussardAndBaggaProtocol.java
@@ -46,7 +46,7 @@ public class BussardAndBaggaProtocol extends DBProtocol{
 	}
 
 	@Override
-	public int getMemory(int n) {
+	public long getMemory(int n) {
 		return getTotalBitsExchanged(n)+2*n;
 	}
