+++ src/protocols/HanckeAndKuhnProtocol.java
@@ -47,7 +47,7 @@ public class HanckeAndKuhnProtocol extends DBProtocol{
 		return false;
 	}
 	@Override
-	public int getMemory(int n) {
+	public long getMemory(int n) {
 		return getTotalBitsExchanged(n)+2*n;
 	}
 	@Override
