+++ src/protocols/MADProtocol.java
@@ -53,7 +53,7 @@ public class MADProtocol extends DBProtocol{
 		return false;
 	}
 	@Override
-	public int getMemory(int n) {
+	public long getMemory(int n) {
 		return getTotalBitsExchanged(n)+2*n;
 	}
 	@Override
